requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
pyjwt==2.8.0
bcrypt==4.0.1
bson==0.5.10
numpy==1.24.3
scikit-learn==1.3.0